Detach ephemeral workers from the import route
