+++ src/kitaru/server/domain/names.py
@@ -43,7 +43,6 @@ VERSION_ALLOWED_SEPARATORS = frozenset({"-", "_", ".", "+", "/"})
 RESERVED_NAMESPACE = "kitaru"
 
 NAMESPACE_SEPARATOR = "/"
-RESERVED_LABEL_PREFIX = f"{RESERVED_NAMESPACE}{NAMESPACE_SEPARATOR}"
 
 MAX_NAME_LENGTH = 255
