+++ src/kitaru/server/domain/task.py
@@ -82,6 +82,7 @@ CONTRACT_ENV_NAMES = frozenset(
     {"KITARU_API_URL", "KITARU_API_KEY", "KITARU_API_TOKEN", "KITARU_REPLAY_ID"}
 )
 CONTRACT_ENV_PREFIX = "KITARU_TASK_"
+RESERVED_LABEL_PREFIX = "kitaru/"
 
 
 class TaskNotFound(NotFoundError):
